+++ static/components/u-views.js
@@ -5,19 +5,32 @@ customElements.define(
   "u-views",
   class Views extends Component {
     init() {
-      const { views, target } = this.dataset;
+      const { views, target, state } = this.dataset;
       this.classList.add("tools");
       const targetElement = document.getElementById(target);
+      const group = push(this, "div", { class: "tools-group" });
 
       if (!targetElement || !views) {
         return;
+      }
+
+      if (initialViews.size === 0 && state) {
+        for (const view of JSON.parse(state)) {
+          initialViews.add(view);
+        }
+
+        history.replaceState(
+          null,
+          "",
+          `#${Array.from(initialViews).join(",")}`,
+        );
       }
 
       const visibleViews = new Set();
 
       for (const view of JSON.parse(views)) {
         const visible = initialViews.has(view);
-        const button = push(this, "button", {
+        const button = push(group, "button", {
           type: "button",
           class: "button is-secondary",
           "aria-pressed": String(visible),
@@ -42,7 +55,11 @@ customElements.define(
       });
 
       const visible = Array.from(views).join(",");
-      history.replaceState(null, "", `#${visible}`);
+      history.replaceState(
+        null,
+        "",
+        visible ? `#${visible}` : location.pathname,
+      );
     }
   },
 );